requirements.txt
gunicorn==20.1.0
albumentations==1.1.0
pyyaml>=5.1
webcolors
tensorboard
tqdm
ensemble-boxes
timm>=0.5.4
omegaconf
pycocotools
gdown>=4.4.0
flask-cors
flask_ngrok
cryptography
tabulate
segmentation-models-pytorch
opencv-python-headless==4.7.0.72
ultralytics
numba>=0.49.0 # not directly required, pinned by Snyk to avoid a vulnerability
rsa>=4.7 # not directly required, pinned by Snyk to avoid a vulnerability
flask_sqlalchemy
mysql-connector-python
Flask-Migrate
Werkzeug==2.1.1